+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v2026.2.0 - 2026-2-21
+
+- Adds document symbol features. See #117. Thanks @Shresht7
+- Updated some dependencies
+
+## v2025.7.0 - 2025-7-11
+
+- Fixes ESPHome local failing into loop trying to load and causing high CPU usage. 🐛
+- Uses Python extension to figure out the "Selected interpreter" it will try to run esphome in the
+  selected virtual environment. Need to restart vscode after changing the venv. 🐍
+- Pulling dev schemas from cloud ☁️ up to once every 6 hours.
+- Fixed tests 🧪
+
 ## v2025.6.0 - 2025-6-16
 
 - Fixes YAML errors not reporting on right location